+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Show better error message for undefined component field
 - Rename a `only-parse` (`-p`) option to `only-resolved` (`-r`) in `domain:check` command
 - Add domain type for `Handler`
+- Allow to use FQ name for events and use deeper event name
 
 ## 0.1.0 - 2020-08-12
 - Initial implementation